- Add explicit casting for reading NED parameters - Use uint64_t instead of unsigned long for pseudonyms - New NNPDA library to match the new pseudonym type and compilied for Windows as well

+++ Prext/base/BasePrivLayer.cc
@@ -41,7 +41,7 @@ void BasePrivLayer::initialize(int stage)
 
 }
 void BasePrivLayer::finish() {
-    emit(sPsynmchng, (unsigned long)psynmCnt-1); //psynmCnt is initialized by 1
+    emit(sPsynmchng, (unsigned int)psynmCnt-1); //psynmCnt is initialized by 1
     emit(sLifetime, (simTime() - startTime).dbl());
     emit(sPtimes, (simTime() - psynmStartTime).dbl());
     recordParametersAsScalars();
